backend_byte/requirements.txt
Flask==2.2.5
Flask-SQLAlchemy==3.1.1
Flask-Migrate==4.0.5
Flask-JWT-Extended==4.5.2
Flask-Cors==3.0.10
Flask-Smorest==0.41.0
apispec==6.3.0
python-dotenv==1.0.0
mysql-connector-python==8.0.33
Werkzeug==2.2.2
